+++ RELEASENOTES.md
@@ -58,3 +58,4 @@
 - Fix issue with looping sounds being removed entirely when out of range that prevented them from being added when back in range.
 - Fix reading directory as file support with -file parameter.
 - Fix invulnerability and sector colormaps to match Doom I + II behavior.
+- Fix crash in sound management when adding in waiting looping sounds.